Joint probability functions

Computing a probability function and its gradient

JointChance.compute_with_Genz — Function

```julia
compute_with_Genz(j, κ, Σ, μ, size, rng)
```

This function computes a separable joint probability function and its gradient using 
Quasi-Monte Carlo. The function is defined as

```math
f(x) = h(x(t) ≥ ξ(t) ∀t = j, ..., j+κ).
```

The function takes 6 arguments:
* `j` - index in Σ and μ from which the joint probability applies.
* `κ` - duration of an outage
* `Σ` - covariance matrix
* `μ` - mean vector
* `size` - sample
* `rng` - random number generator

If `size` and `rng` are not specified, they take the default values `size = 5000` and `rng = MersenneTwister(1234)`.
With the latter, we fix the random number generator `rng` to guarantee the stability of the gradients in the solver iterations.
This tempers with the precision of the solution found, yet it is necessary for solver convergence.
# TODO formulate better here

JointChance.compute_with_SRD — Function

```julia
compute_with_SRD(j, κ, Σ, μ, size, rng)
```

This function computes a separable joint probability function and its gradient using 
Spherical Radial Decomposition. The function is defined as

```math
f(x) = h(x(t) ≥ ξ(t) ∀t = j, ..., j+κ).
```

The function takes 6 arguments:
* `w` - index in Σ and μ from which the joint probability applies.
* `κ` - duration of an outage
* `Σ` - covariance matrix
* `μ` - mean vector
* `size` - sample
* `rng` - random number generator

If `size` and `rng` are not specified, they take the default values `size = 5000` and `rng = MersenneTwister(1234)`.
With the latter, we fix the random number generator `rng` to guarantee the stability of the gradients in the solver iterations.
This tempers with the precision of the solution found, yet it is necessary for solver convergence.
# TODO formulate better here

Adding the probablity functions as constraints to an existing JuMP model

JointChance.add_constraints_Genz — Function

add_constraints_Genz(m, x, idx, κ, Σ, μ, p)

JointChance.add_constraints_SRD — Function

add_constraints_SRD(m, x, idx, κ, Σ, μ, p)